+++ .gitignore
@@ -1,3 +1,4 @@
 /vendor
 /node_modules
 .idea
+/.idea
