--- app/Events/DelaySendProductKey.php
@@ -1,34 +1,0 @@
-<?php
-
-namespace App\Events;
-
-use App\Events\Event;
-use Illuminate\Queue\SerializesModels;
-use Illuminate\Contracts\Broadcasting\ShouldBroadcast;
-
-class DelaySendProductKey extends Event
-{
-    use SerializesModels;
-
-    public $order;
-
-    /**
-     * Create a new event instance.
-     *
-     * @return void
-     */
-    public function __construct($order)
-    {
-        $this->order = $order;
-    }
-
-    /**
-     * Get the channels the event should be broadcast on.
-     *
-     * @return array
-     */
-    public function broadcastOn()
-    {
-        return [];
-    }
-}
