+++ app/Jobs/SendDelayProductKeyEmail.php
@@ -7,20 +7,22 @@ use Illuminate\Queue\SerializesModels;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Contracts\Bus\SelfHandling;
 use Illuminate\Contracts\Queue\ShouldQueue;
-use App\Order;
-use App\ProductKey;
+use App\Jobs\SendProductKeyEmail;
+use Illuminate\Foundation\Bus\DispatchesJobs;
 
 class SendDelayProductKeyEmail extends Job implements SelfHandling, ShouldQueue
 {
     use InteractsWithQueue, SerializesModels;
-    protected $order;
+    use DispatchesJobs;
+
+    public $order;
 
     /**
      * Create a new job instance.
      *
      * @return void
      */
-    public function __construct(Order $order)
+    public function __construct($order)
     {
         $this->order = $order;
     }
@@ -32,6 +34,7 @@ class SendDelayProductKeyEmail extends Job implements SelfHandling, ShouldQueue
      */
     public function handle()
     {
-        //
+        $job = (new SendProductKeyEmail($this->order))->delay(300);
+        $this->dispatch($job);
     }
 }
